REFACTOR: (SPOT-166) Route Util 클래스를 Event 도메인 하위로 옮긴다
