+++ src/main/java/com/meetup/server/event/dto/response/EventStartPointResponse.java
@@ -2,7 +2,7 @@ package com.meetup.server.event.dto.response;
 
 import com.meetup.server.event.domain.Event;
 import com.meetup.server.startpoint.domain.StartPoint;
-import com.meetup.server.startpoint.util.UsernameExtractor;
+import com.meetup.server.event.util.UsernameExtractor;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
