+++ src/main/java/com/meetup/server/event/dto/response/route/RouteResponse.java
@@ -3,8 +3,8 @@ package com.meetup.server.event.dto.response.route;
 import com.meetup.server.global.clients.kakao.mobility.KakaoMobilityResponse;
 import com.meetup.server.global.clients.odsay.OdsayTransitRouteSearchResponse;
 import com.meetup.server.startpoint.domain.StartPoint;
-import com.meetup.server.startpoint.util.AddressConverter;
-import com.meetup.server.startpoint.util.UsernameExtractor;
+import com.meetup.server.event.util.AddressConverter;
+import com.meetup.server.event.util.UsernameExtractor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
