+++ src/main/java/com/meetup/server/event/implement/route/RouteAssembler.java
@@ -8,7 +8,7 @@ import com.meetup.server.global.clients.odsay.OdsayTransitRouteSearchResponse;
 import com.meetup.server.parkinglot.implement.ParkingLotFinder;
 import com.meetup.server.parkinglot.infrastructure.jpa.projection.ClosestParkingLot;
 import com.meetup.server.startpoint.domain.StartPoint;
-import com.meetup.server.startpoint.util.RouteExtractor;
+import com.meetup.server.event.util.RouteExtractor;
 import com.meetup.server.subway.domain.Subway;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
