+++ src/main/java/com/meetup/server/event/util/RouteExtractor.java
@@ -1,8 +1,11 @@
-package com.meetup.server.startpoint.util;
+package com.meetup.server.event.util;
 
 import com.meetup.server.event.dto.response.route.DrivingInfoResponse;
 import com.meetup.server.global.clients.odsay.OdsayTransitRouteSearchResponse;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class RouteExtractor {
 
     public static int extractValidTransitTotalTime(OdsayTransitRouteSearchResponse transitResponse) {
